修复了某些情况下导入失败的问题

+++ eoa-lib/network-html-api/src/commonMain/kotlin/top/kagg886/sylu_eoa/api/html/EOAHTMLClient.kt
@@ -454,7 +454,7 @@ internal class EOAHTMLClient : EOAClient {
                 i.rangeEveryDay.map { lessonNumber ->
                     val time = getTimeByLessonNumber(lessonNumber)
                     ClassTable(
-                        id = i.id.toLong(),
+                        id = i.id,
                         name = i.name,
                         teacher = i.teacher,
                         room = i.room,
